requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
azure-ai-translation-document==1.1.0
azure-ai-translation-text==1.0.0b1
azure-core==1.32.0
azure-storage-blob==12.24.1
blinker==1.9.0
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.1
distro==1.9.0
dnspython==2.7.0
gitdb==4.0.12
GitPython==3.1.44
groq==0.18.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
isodate==0.7.2
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lxml==5.3.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
msrest==0.7.1
narwhals==1.26.0
numpy==2.2.2
oauthlib==3.2.2
packaging==24.2
pandas==2.2.3
pillow==11.1.0
protobuf==5.29.3
pyarrow==19.0.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
pymongo==4.11.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
render==1.0.0
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.22.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.42.0
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
watchdog==6.0.0
